Extensively refactor submission unpacking
The core of this refactoring centered around minimizing the size of the code
associated with submission unpacking via applying a more updated software
engineering philosphy which isn't so allergic to somewhat longer functions.

Additionally, I discovered that the .txt file added by BBLearn to each
submission contains all of the information necessary to do everything that this
code does without ever having to explore the file system, so a new refactoring
should happen again later.